don't try and publish none existent test results
there currently are no tests so trying to publish the results causes the jenkins build to fail as it thinks there is a misconfiguration. For now we just comment out the junit publishing action. Hopefully at some future point there might be some tests at which point we can put it back.